Fixed BottomBar and CustomHeader and more. Added ProfileScreen

+++ domain/src/main/java/com/example/domain/ui/presentation/feature/favorite/model/Favorite.kt
@@ -7,7 +7,14 @@ import kotlinx.serialization.Serializable
 data class Favorite(
     val id: Long,
     val userId: Long,
-    val productId: Int,
+    val productId: Long,
     val createdAt: String,
     val product: Product
 )
+//
+//data class FavoriteProduct(
+//    val id : Long,
+//    val name: String,
+//    val images: String,
+//    val price: Double
+//)
